Load in apple image and draw heart symbol for lives.

+++ myLib.h
@@ -167,7 +167,7 @@ void waitForVblank();
 MOVOBJ movobj_new(int row, int col, int rvel, int cvel, int size, u16 color);
 Game game_new(enum GameState state);
 
-void drawImage3(int r, int c, int width, int height, const int* image);
+void drawImage3(int r, int c, int width, int height,const unsigned short* image);
 void drawRect(int r, int c, int height, int width, unsigned short color);
 void fillScreen(unsigned short color);
 void createShot(Game* game, int row, int col, int up);
